requirements.txt
numpy
json
moviepy
nemo-toolkit
torchaudio
soundfile
scipy
temp
whisperx